+++ hub-client/src/locales/nl.ts
@@ -49,6 +49,9 @@ const nl = {
 		hub_homepage_welcome_auth: 'Welkom bij de {0}-hub.',
 		welcome: 'Welkom',
 		highlighted_hubs: 'Uitgelichte Hubs',
+		welcome_at: 'Welkom bij',
+		discover_hubs: 'Ontdek Hubs',
+		hub_card_about: 'Over',
 	},
 	register: {
 		register: 'Registreren',
@@ -130,6 +133,8 @@ const nl = {
 	others: {
 		nop: 'Deze functionaliteit is nog niet gerealiseerd.',
 		search: 'Zoeken',
+		search_hubs: 'Zoek naar Hubs',
+		search_hubs_not_found: 'Geen Hubs gevonden',
 		search_room: 'Zoek in dit gesprek',
 		searching: 'Bezig met zoeken...',
 		search_nothing_found: 'Niets gevonden',