+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Use the [MIGRATE] and [BREAKING] prefixes together with another one if that make
 - [BUG] Right padding of menu-items was off, weird tooltips when hovering over room in menuitem.
 - [BUG] Dialog modal on Safari browser did not cover whole screen
 - [BUG] Styling of Hub Icon was wrong
+- [NEW] Updated the design for the discover hub & login page
+- [NEW] Update input bar design & make hotfixes for the button opacity and placeholder text
 
 ## 6 February 2025 - v1.0.5
 
@@ -30,7 +32,6 @@ Use the [MIGRATE] and [BREAKING] prefixes together with another one if that make
 - Hub admins can now change the hub icon from the hub settings dialog.
 - Removed legacy static pages and some of the underlying code, see #1035.
 - Hubs are now identified by their hubId in /bar/state and hub access tokens are stored in /bar/state.
-- Update input bar design & make hotfixes for the button opacity and placeholder text
 
 ## 21 January 2025 - v1.0.4